+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Configure Steep on this gem
+
 ## [0.1.0] - 2024-05-10
 
 - Initial release